// src/gql/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: { input: unknown; output: unknown; }
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: { input: unknown; output: unknown; }
  /** The `Upload` scalar type represents a file upload. */
  Upload: { input: unknown; output: unknown; }
};
// ...
export type CategoryRelationResponseCollection = {
  data: Array<CategoryEntity>;
};
// ...
export type CollectionEntityResponse = {
  data?: Maybe<CollectionEntity>;
};
// ...
export type Product = {
  categories?: Maybe<CategoryRelationResponseCollection>;
  collection?: Maybe<CollectionEntityResponse>;
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  description: Scalars['String']['output'];
  images: UploadFileRelationResponseCollection;
  name: Scalars['String']['output'];
  price: Scalars['Float']['output'];
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  slug: Scalars['String']['output'];
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
};
// ...
export type ProductEntity = {
  attributes?: Maybe<Product>;
  id?: Maybe<Scalars['ID']['output']>;
};
// ...
export type UploadFileRelationResponseCollection = {
  data: Array<UploadFileEntity>;
};
// ...
export type ProductGetByIdQueryVariables = Exact<{
  id: Scalars['ID']['input'];
}>;


export type ProductGetByIdQuery = { product?: { data?: { id?: string | null, attributes?: { name: string, description: string, price: number, slug: string, images: { data: Array<{ attributes?: { url: string, alternativeText?: string | null } | null }> }, categories?: { data: Array<{ id?: string | null, attributes?: { name: string, slug: string } | null }> } | null } | null } | null } | null };

export type ProductsGetListQueryVariables = Exact<{
  pageSize?: InputMaybe<Scalars['Int']['input']>;
  page?: InputMaybe<Scalars['Int']['input']>;
}>;


export type ProductsGetListQuery = { products?: { data: Array<{ id?: string | null, attributes?: { name: string, description: string, price: number, slug: string, images: { data: Array<{ attributes?: { url: string, alternativeText?: string | null } | null }> }, categories?: { data: Array<{ id?: string | null, attributes?: { name: string, slug: string } | null }> } | null } | null }>, meta: { pagination: { pageCount: number } } } | null };

export type ProductsGetListByCategorySlugQueryVariables = Exact<{
  pageSize?: InputMaybe<Scalars['Int']['input']>;
  page?: InputMaybe<Scalars['Int']['input']>;
  categorySlug: Scalars['String']['input'];
}>;


export type ProductsGetListByCategorySlugQuery = { products?: { data: Array<{ id?: string | null, attributes?: { name: string, description: string, price: number, slug: string, images: { data: Array<{ attributes?: { url: string, alternativeText?: string | null } | null }> }, categories?: { data: Array<{ id?: string | null, attributes?: { name: string, slug: string } | null }> } | null } | null }>, meta: { pagination: { pageCount: number } } } | null };

export type ProductsGetListByNameQueryVariables = Exact<{
  name?: InputMaybe<Scalars['String']['input']>;
}>;


export type ProductsGetListByNameQuery = { products?: { data: Array<{ id?: string | null, attributes?: { name: string, description: string, price: number, slug: string, images: { data: Array<{ attributes?: { url: string, alternativeText?: string | null } | null }> }, categories?: { data: Array<{ id?: string | null, attributes?: { name: string, slug: string } | null }> } | null } | null }>, meta: { pagination: { total: number } } } | null };

export class TypedDocumentString<TResult, TVariables>
  extends String
  implements DocumentTypeDecoration<TResult, TVariables>
{
  __apiType?: DocumentTypeDecoration<TResult, TVariables>['__apiType'];

  constructor(private value: string, public __meta__?: Record<string, any>) {
    super(value);
  }

  toString(): string & DocumentTypeDecoration<TResult, TVariables> {
    return this.value;
  }
}
// ...
export const ProductGetByIdDocument = new TypedDocumentString(`
    query ProductGetById($id: ID!) {
  product(id: $id) {
    data {
      ...ProductListItem
    }
  }
}
    fragment ProductListItem on ProductEntity {
  id
  attributes {
    ...ProductListItemAttributes
  }
}
fragment ProductListItemAttributes on Product {
  name
  description
  price
  images {
    data {
      attributes {
        url
        alternativeText
      }
    }
  }
  categories {
    data {
      id
      attributes {
        name
        slug
      }
    }
  }
  slug
}`) as unknown as TypedDocumentString<ProductGetByIdQuery, ProductGetByIdQueryVariables>;
export const ProductsGetListDocument = new TypedDocumentString(`
    query ProductsGetList($pageSize: Int = 4, $page: Int = 1) {
  products(pagination: {pageSize: $pageSize, page: $page}) {
    data {
      ...ProductListItem
    }
    meta {
      pagination {
        pageCount
      }
    }
  }
}
    fragment ProductListItem on ProductEntity {
  id
  attributes {
    ...ProductListItemAttributes
  }
}
fragment ProductListItemAttributes on Product {
  name
  description
  price
  images {
    data {
      attributes {
        url
        alternativeText
      }
    }
  }
  categories {
    data {
      id
      attributes {
        name
        slug
      }
    }
  }
  slug
}`) as unknown as TypedDocumentString<ProductsGetListQuery, ProductsGetListQueryVariables>;
export const ProductsGetListByCategorySlugDocument = new TypedDocumentString(`
    query ProductsGetListByCategorySlug($pageSize: Int = 4, $page: Int = 1, $categorySlug: String!) {
  products(
    pagination: {pageSize: $pageSize, page: $page}
    filters: {categories: {slug: {eq: $categorySlug}}}
  ) {
    data {
      ...ProductListItem
    }
    meta {
      pagination {
        pageCount
      }
    }
  }
}
    fragment ProductListItem on ProductEntity {
  id
  attributes {
    ...ProductListItemAttributes
  }
}
fragment ProductListItemAttributes on Product {
  name
  description
  price
  images {
    data {
      attributes {
        url
        alternativeText
      }
    }
  }
  categories {
    data {
      id
      attributes {
        name
        slug
      }
    }
  }
  slug
}`) as unknown as TypedDocumentString<ProductsGetListByCategorySlugQuery, ProductsGetListByCategorySlugQueryVariables>;
export const ProductsGetListByNameDocument = new TypedDocumentString(`
    query ProductsGetListByName($name: String) {
  products(filters: {name: {containsi: $name}}) {
    data {
      ...ProductListItem
    }
    meta {
      pagination {
        total
      }
    }
  }
}
    fragment ProductListItem on ProductEntity {
  id
  attributes {
    ...ProductListItemAttributes
  }
}
fragment ProductListItemAttributes on Product {
  name
  description
  price
  images {
    data {
      attributes {
        url
        alternativeText
      }
    }
  }
  categories {
    data {
      id
      attributes {
        name
        slug
      }
    }
  }
  slug
}`) as unknown as TypedDocumentString<ProductsGetListByNameQuery, ProductsGetListByNameQueryVariables>;